Show result view

+++ Assets/Scripts/Models/StatsModel.cs
@@ -5,7 +5,7 @@ namespace Models
 {
     public class StatsModel : IDisposable
     {
-        public ReactiveProperty<float> KilledCubes = new(0);
+        public ReactiveProperty<int> KilledCubes = new(0);
 
         public void Dispose()
         {
